refactor launchers and test code
