+++ package/kedro_viz/server.py
@@ -1,6 +1,5 @@
 """`kedro_viz.server` provides utilities to launch a webserver
 for Kedro pipeline visualisation."""
-import webbrowser
 from pathlib import Path
 from typing import Any, Dict, Optional
 
@@ -20,11 +19,6 @@ from kedro_viz.integrations.kedro import data_loader as kedro_data_loader
 from kedro_viz.integrations.kedro.sqlite_store import SQLiteStore
 
 DEV_PORT = 4142
-
-
-def is_localhost(host) -> bool:
-    """Check whether a host is a localhost"""
-    return host in ("127.0.0.1", "localhost", "0.0.0.0")
 
 
 def populate_data(
@@ -54,7 +48,6 @@ def populate_data(
 def run_server(
     host: str = DEFAULT_HOST,
     port: int = DEFAULT_PORT,
-    browser: Optional[bool] = None,
     load_file: Optional[str] = None,
     save_file: Optional[str] = None,
     pipeline_name: Optional[str] = None,
@@ -69,7 +62,6 @@ def run_server(
     Args:
         host: the host to launch the webserver
         port: the port to launch the webserver
-        browser: whether to open the default browser automatically
         load_file: if a valid JSON containing API response data is provided,
             the API of the server is created from the JSON.
         save_file: if provided, the data returned by the API will be saved to a file.
@@ -84,6 +76,7 @@ def run_server(
             take precedence over) the parameters retrieved from the project
             configuration.
     """
+    print("Starting Kedro Viz Backend Server...")
     if load_file is None:
         path = Path(project_path) if project_path else Path.cwd()
         catalog, pipelines, session_store, stats_dict = kedro_data_loader.load_data(
