+++ package/tests/test_launchers/test_utils.py
@@ -1,0 +1,28 @@
+from unittest import mock
+
+import pytest
+
+from kedro_viz.launchers.utils import start_browser
+
+
+@pytest.mark.parametrize(
+    "ip,should_browser_open",
+    [
+        ("0.0.0.0", True),
+        ("127.0.0.1", True),
+        ("localhost", True),
+        ("8.8.8.8", False),
+    ],
+)
+@mock.patch("kedro_viz.launchers.utils.webbrowser")
+def test_browser_open(
+    webbrowser,
+    ip,
+    should_browser_open,
+    mocker,
+):
+    start_browser(ip, port=4141)
+    if should_browser_open:
+        webbrowser.open_new.assert_called_once()
+    else:
+        webbrowser.open_new.assert_not_called()
